# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ProjectSend",
        "version": "1.0.0",
        "description": "Staff-authenticated REST API for ProjectSend. Authenticate with a bearer token created under Settings \u2192 API tokens. See docs/api-guide.md for the narrative guide."
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/comments/pending": {
            "get": {
                "operationId": "comments.pending",
                "description": "Scoped by the same library boundary as everything else: a\nclient-scoped token sees pending comments only on files its owner\ncould already open. Oldest first, so working through the list means\nworking through the backlog.\n\nRequires a token with the ability: `moderate_comments`.",
                "summary": "List comments awaiting approval",
                "tags": [
                    "CommentModeration"
                ],
                "responses": {
                    "200": {
                        "description": "Array of `FileCommentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FileCommentResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "FileCommentResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "file_id": {
                        "type": "integer"
                    },
                    "body": {
                        "type": "string"
                    },
                    "visibility": {
                        "type": "string"
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": [
                                    "integer",
                                    "null"
                                ]
                            },
                            "name": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "guest",
                                    "staff",
                                    "client"
                                ]
                            }
                        },
                        "required": [
                            "id",
                            "name",
                            "type"
                        ]
                    },
                    "conversation": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Whose conversation this is, when it is one client's rather\nthan every client's. Staff only, and null on a comment\naddressed to all of them.",
                        "properties": {
                            "client_id": {
                                "type": [
                                    "integer",
                                    "null"
                                ]
                            },
                            "client_name": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "client_id",
                            "client_name"
                        ]
                    },
                    "approved": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "edited_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "file_id",
                    "body",
                    "visibility",
                    "author",
                    "conversation",
                    "approved",
                    "created_at",
                    "edited_at"
                ],
                "title": "FileCommentResource"
            },
...
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}